+++ src/index.ts
@@ -151,9 +151,9 @@ registerComponentController('slider', Slider);
 // 注册 Interaction Action
 import { registerAction } from './core';
 import ActiveRegion from './interaction/action/active-region';
-import SiblingTooltip from './interaction/action/component/sibling-tooltp';
+import SiblingTooltip from './interaction/action/component/tooltip/sibling';
 import TooltipAction from './interaction/action/component/tooltip/geometry';
-import ComponentTooltipAction from './interaction/action/component/tooltip/component';
+import EllipsisTextAction from './interaction/action/component/tooltip/ellipsis-text';
 
 import ElmentActive from './interaction/action/element/active';
 import ElementLinkByColor from './interaction/action/element/link-by-color';
@@ -197,7 +197,7 @@ import ScaleZoom from './interaction/action/view/scale-zoom';
 
 registerAction('tooltip', TooltipAction);
 registerAction('sibling-tooltip', SiblingTooltip);
-registerAction('component-tooltip', ComponentTooltipAction);
+registerAction('ellipsis-text', EllipsisTextAction);
 registerAction('element-active', ElmentActive);
 registerAction('element-single-active', ElmentSingleActive);
 registerAction('element-range-active', ElmentRangeActive);
@@ -283,18 +283,18 @@ registerInteraction('tooltip', {
   ],
 });
 
-registerInteraction('component-tooltip', {
+registerInteraction('ellipsis-text', {
   start: [
-    { trigger: 'legend-item-name:mousemove', action: 'component-tooltip:show', throttle: { wait: 50, leading: true, trailing: false } },
-    { trigger: 'legend-item-name:touchstart', action: 'component-tooltip:show', throttle: { wait: 50, leading: true, trailing: false } },
-    { trigger: 'axis-label:mousemove', action: 'component-tooltip:show', throttle: { wait: 50, leading: true, trailing: false } },
-    { trigger: 'axis-label:touchstart', action: 'component-tooltip:show', throttle: { wait: 50, leading: true, trailing: false } },
+    { trigger: 'legend-item-name:mousemove', action: 'ellipsis-text:show', throttle: { wait: 50, leading: true, trailing: false } },
+    { trigger: 'legend-item-name:touchstart', action: 'ellipsis-text:show', throttle: { wait: 50, leading: true, trailing: false } },
+    { trigger: 'axis-label:mousemove', action: 'ellipsis-text:show', throttle: { wait: 50, leading: true, trailing: false } },
+    { trigger: 'axis-label:touchstart', action: 'ellipsis-text:show', throttle: { wait: 50, leading: true, trailing: false } },
   ],
   end: [
-    { trigger: 'legend-item-name:mouseleave', action: 'component-tooltip:hide' },
-    { trigger: 'legend-item-name:touchend', action: 'component-tooltip:hide' },
-    { trigger: 'axis-label:mouseleave', action: 'component-tooltip:hide' },
-    { trigger: 'axis-label:touchend', action: 'component-tooltip:hide' },
+    { trigger: 'legend-item-name:mouseleave', action: 'ellipsis-text:hide' },
+    { trigger: 'legend-item-name:touchend', action: 'ellipsis-text:hide' },
+    { trigger: 'axis-label:mouseleave', action: 'ellipsis-text:hide' },
+    { trigger: 'axis-label:touchend', action: 'ellipsis-text:hide' },
   ],
 });
 
@@ -647,6 +647,7 @@ export * from './core';
 // 一些工具方法导出
 import { getAngle, polarToCartesian } from './util/graphics';
 import { rotate, transform, translate, zoom } from './util/transform';
+import EllipsisText from './interaction/action/component/tooltip/ellipsis-text';
 export const Util = {
   translate,
   rotate,
