+++ src/interaction/action/component/tooltip/sibling.ts
@@ -1,8 +1,8 @@
 import { each } from '@antv/util';
-import { View } from '../../../chart';
-import { Point } from '../../../interface';
-import { getSiblingPoint, getSilbings } from '../util';
-import TooltipAction from './tooltip/geometry';
+import { View } from '../../../../chart';
+import { Point } from '../../../../interface';
+import { getSiblingPoint, getSilbings } from '../../util';
+import TooltipAction from './geometry';
 
 /**
  * 存在多个 view 时，控制其他 view 上的 tooltip 显示
